[serve] Update API stability annotations for 2.7 release (#39126)
Updates/cleans up API stability annotations in anticipation of the 2.7 release. Changes aside from upgrading stabilities:

- Deprecate `serve.build` and `BuiltApplication`.
- Move `DeploymentConfig` and `ReplicaConfig` to a private file without an annotation (they're not publicly-exposed APIs).
- `s/get_global_client/_get_global_client` (also not a publicly-exposed API).
- Remove annotation from `RequestContext` and make it `_RequestContext` (also not a publicly-exposed API).
- Add `ProxyLocation` type to supersede `DeploymentMode`.
- Add docstrings to `HTTPOptions` and `ProxyLocation`.
