+++ dashboard/client/src/pages/serve/ServeApplicationsListPage.component.test.tsx
@@ -4,7 +4,7 @@ import { getActor } from "../../service/actor";
 import { getServeApplications } from "../../service/serve";
 import {
   ServeApplicationStatus,
-  ServeDeploymentMode,
+  ServeProxyLocation,
   ServeSystemActorStatus,
 } from "../../type/serve";
 import { TEST_APP_WRAPPER } from "../../util/test-utils";
@@ -45,7 +45,7 @@ describe("ServeApplicationsListPage", () => {
           node_id: "node:12345",
           actor_id: "actor:12345",
         },
-        proxy_location: ServeDeploymentMode.EveryNode,
+        proxy_location: ServeProxyLocation.EveryNode,
         applications: {
           home: {
             name: "home",
