+++ dashboard/client/src/type/serve.ts
@@ -78,8 +78,9 @@ export type ServeReplica = {
   log_file_path: string | null;
 };
 
-// Keep in sync with DeploymentMode in python/ray/serve/config.py
-export enum ServeDeploymentMode {
+// Keep in sync with ProxyLocation in python/ray/serve/config.py
+export enum ServeProxyLocation {
+  Disabled = "Disabled",
   NoServer = "NoServer",
   HeadOnly = "HeadOnly",
   EveryNode = "EveryNode",
@@ -117,7 +118,7 @@ export type ServeApplicationsRsp = {
   grpc_options: {
     port: number;
   };
-  proxy_location: ServeDeploymentMode;
+  proxy_location: ServeProxyLocation;
   controller_info: ServeSystemActor;
   proxies: {
     [name: string]: ServeHttpProxy;
