+++ doc/source/serve/api/index.md
@@ -44,19 +44,6 @@ This is fixed by added custom filename mappings in `source/conf.py` (look for "a
    serve.handle.RayServeSyncHandle
 ```
 
-#### Advanced APIs
-
-```{eval-rst}
-.. autosummary::
-   :nosignatures:
-   :toctree: doc/
-
-   serve.get_replica_context
-   serve.get_multiplexed_model_id
-   serve.get_app_handle
-   serve.get_deployment_handle
-```
-
 ### Running Applications
 
 ```{eval-rst}
@@ -66,9 +53,34 @@ This is fixed by added custom filename mappings in `source/conf.py` (look for "a
 
    serve.run
    serve.delete
-   serve.start
-   serve.shutdown
    serve.status
+```
+
+### Configuring System-level Options
+
+```{eval-rst}
+.. autosummary::
+   :nosignatures:
+   :toctree: doc/
+
+   serve.start
+   serve.config.ProxyLocation
+   serve.config.HTTPOptions
+   serve.shutdown
+```
+
+#### Advanced APIs
+
+```{eval-rst}
+.. autosummary::
+   :nosignatures:
+   :toctree: doc/
+
+   serve.get_replica_context
+   serve.context.ReplicaContext
+   serve.get_multiplexed_model_id
+   serve.get_app_handle
+   serve.get_deployment_handle
 ```
 
 (serve-cli)=
