+++ python/ray/serve/_private/http_proxy.py
@@ -26,7 +26,6 @@ from ray.serve._private.usage import ServeUsageTag
 from ray._private.utils import get_or_create_event_loop
 from ray._raylet import StreamingObjectRefGenerator
 
-from ray import serve
 from ray.serve.handle import (
     DeploymentHandle,
     _DeploymentResponseBase,
@@ -169,10 +168,14 @@ class GenericProxy(ABC):
 
         self._node_id = node_id
 
-        # Set the controller name so that serve will connect to the
+        # Set the controller name so that serve connects to the
         # controller instance this proxy is running in.
         ray.serve.context._set_internal_replica_context(
-            None, None, controller_name, None, None
+            app_name=None,
+            deployment=None,
+            replica_tag=None,
+            servable_object=None,
+            controller_name=controller_name,
         )
 
         # Used only for displaying the route table.
@@ -188,7 +191,11 @@ class GenericProxy(ABC):
             )
 
         def get_handle(deployment_name, app_name):
-            return serve.context.get_global_client().get_handle(
+            # Delayed import due to circular dependency.
+            # TODO(edoakes): use `get_deployment_handle` public API instead.
+            from ray.serve.context import _get_global_client
+
+            return _get_global_client().get_handle(
                 deployment_name,
                 app_name,
                 sync=False,
@@ -778,7 +785,7 @@ class gRPCProxy(GenericProxy):
             "multiplexed_model_id": multiplexed_model_id,
         }
         ray.serve.context._serve_request_context.set(
-            ray.serve.context.RequestContext(**request_context_info)
+            ray.serve.context._RequestContext(**request_context_info)
         )
         proxy_request.send_request_id(request_id=request_id)
         return handle, request_id
@@ -1266,7 +1273,7 @@ class HTTPProxy(GenericProxy):
                 # "x-request-id" has higher priority than "RAY_SERVE_REQUEST_ID".
                 request_context_info["request_id"] = value.decode()
         ray.serve.context._serve_request_context.set(
-            ray.serve.context.RequestContext(**request_context_info)
+            ray.serve.context._RequestContext(**request_context_info)
         )
         return handle, request_context_info["request_id"]
