+++ python/ray/serve/built_application.py
@@ -6,10 +6,10 @@ from typing import (
 
 from ray.serve.deployment import Deployment
 from ray.serve._private.deploy_utils import get_deploy_args
-from ray.util.annotations import PublicAPI
+from ray.util.annotations import Deprecated
 
 
-@PublicAPI(stability="alpha")
+@Deprecated(message="Only used with `BuiltApplication` which has been deprecated.")
 class ImmutableDeploymentDict(dict):
     def __init__(self, deployments: Dict[str, Deployment]):
         super().__init__()
@@ -23,7 +23,7 @@ class ImmutableDeploymentDict(dict):
         )
 
 
-@PublicAPI(stability="alpha")
+@Deprecated(message="Use the `serve build` CLI command instead.")
 class BuiltApplication:
     """A static, pre-built Serve application.
