+++ python/ray/serve/deployment.py
@@ -10,18 +10,16 @@ from typing import (
     Tuple,
     Union,
 )
+import warnings
 
 from ray.dag.dag_node import DAGNodeBase
 from ray.dag.class_node import ClassNode
 from ray.dag.function_node import FunctionNode
 from ray.util.annotations import Deprecated, PublicAPI
 
-from ray.serve.config import (
-    AutoscalingConfig,
-    DeploymentConfig,
-    ReplicaConfig,
-)
-from ray.serve.context import get_global_client
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
+from ray.serve.config import AutoscalingConfig
+from ray.serve.context import _get_global_client
 from ray.serve.handle import RayServeHandle, RayServeSyncHandle
 from ray.serve.schema import (
     RayActorOptionsSchema,
@@ -39,7 +37,7 @@ from ray.serve._private.utils import (
 logger = logging.getLogger(SERVE_LOGGER_NAME)
 
 
-@PublicAPI(stability="beta")
+@PublicAPI(stability="stable")
 class Application(DAGNodeBase):
     """One or more deployments bound with arguments that can be deployed together.
 
@@ -100,7 +98,7 @@ class Application(DAGNodeBase):
         return getattr(self._get_internal_dag_node(), name)
 
 
-@PublicAPI
+@PublicAPI(stability="stable")
 class Deployment:
     """Class (or function) decorated with the `@serve.deployment` decorator.
 
@@ -237,7 +235,7 @@ class Deployment:
             # this deployment is not exposed over HTTP
             return None
 
-        return get_global_client().root_url + self.route_prefix
+        return _get_global_client().root_url + self.route_prefix
 
     def __call__(self):
         raise RuntimeError(
@@ -245,7 +243,6 @@ class Deployment:
             "Use `deployment.deploy() instead.`"
         )
 
-    @PublicAPI(stability="beta")
     def bind(self, *args, **kwargs) -> Application:
         """Bind the arguments to the deployment and return an Application.
 
@@ -319,7 +316,7 @@ class Deployment:
             max_replicas_per_node=self._replica_config.max_replicas_per_node,
         )
 
-        return get_global_client().deploy(
+        return _get_global_client().deploy(
             self._name,
             replica_config=replica_config,
             deployment_config=self._deployment_config,
@@ -340,7 +337,7 @@ class Deployment:
     def _delete(self):
         """Delete this deployment."""
 
-        return get_global_client().delete_deployments([self._name])
+        return _get_global_client().delete_deployments([self._name])
 
     @guarded_deprecation_warning(instructions=MIGRATION_MESSAGE)
     @Deprecated(message=MIGRATION_MESSAGE)
@@ -377,14 +374,13 @@ class Deployment:
             ServeHandle
         """
 
-        return get_global_client().get_handle(
+        return _get_global_client().get_handle(
             self._name,
             app_name="",
             missing_ok=True,
             sync=sync,
         )
 
-    @PublicAPI
     def options(
         self,
         func_or_class: Optional[Callable] = None,
@@ -541,7 +537,12 @@ class Deployment:
             is_driver_deployment=is_driver_deployment,
         )
 
-    @PublicAPI(stability="alpha")
+    @Deprecated(
+        message=(
+            "This was intended for use with the `serve.build` Python API "
+            "(which has been deprecated). Use `.options()` instead."
+        )
+    )
     def set_options(
         self,
         func_or_class: Optional[Callable] = None,
@@ -572,6 +573,11 @@ class Deployment:
         Refer to the @serve.deployment decorator docstring for all non-private
         arguments.
         """
+        if not _internal:
+            warnings.warn(
+                "`.set_options()` is deprecated. "
+                "Use `.options()` or an application builder function instead."
+            )
 
         validated = self.options(
             func_or_class=func_or_class,
