+++ python/ray/serve/scripts.py
@@ -20,7 +20,7 @@ from ray.autoscaler._private.cli_logger import cli_logger
 from ray.dashboard.modules.dashboard_sdk import parse_runtime_env_args
 from ray.dashboard.modules.serve.sdk import ServeSubmissionClient
 from ray.serve.api import build as build_app
-from ray.serve.config import DeploymentMode, gRPCOptions
+from ray.serve.config import DeploymentMode, ProxyLocation, gRPCOptions
 from ray.serve._private.constants import (
     DEFAULT_GRPC_PORT,
     DEFAULT_HTTP_HOST,
@@ -165,9 +165,9 @@ def cli():
 )
 @click.option(
     "--proxy-location",
-    default=DeploymentMode.EveryNode,
+    default=ProxyLocation.EveryNode,
     required=False,
-    type=click.Choice(list(DeploymentMode)),
+    type=click.Choice(list(ProxyLocation)),
     help="Location of the proxies. Defaults to EveryNode.",
 )
 @click.option(
