+++ python/ray/serve/tests/test_config.py
@@ -3,11 +3,11 @@ from pydantic import ValidationError
 
 from ray import cloudpickle
 
+from ray.serve._private.config import DeploymentConfig, ReplicaConfig
 from ray.serve.config import (
-    DeploymentConfig,
     DeploymentMode,
     HTTPOptions,
-    ReplicaConfig,
+    ProxyLocation,
     gRPCOptions,
 )
 from ray.serve.config import AutoscalingConfig
@@ -496,6 +496,31 @@ def test_grpc_options():
     assert grpc_options.grpc_servicer_func_callable == [
         add_UserDefinedServiceServicer_to_server
     ]
+
+
+def test_proxy_location_to_deployment_mode():
+    assert (
+        ProxyLocation._to_deployment_mode(ProxyLocation.Disabled)
+        == DeploymentMode.NoServer
+    )
+    assert (
+        ProxyLocation._to_deployment_mode(ProxyLocation.HeadOnly)
+        == DeploymentMode.HeadOnly
+    )
+    assert (
+        ProxyLocation._to_deployment_mode(ProxyLocation.EveryNode)
+        == DeploymentMode.EveryNode
+    )
+
+    assert ProxyLocation._to_deployment_mode("Disabled") == DeploymentMode.NoServer
+    assert ProxyLocation._to_deployment_mode("HeadOnly") == DeploymentMode.HeadOnly
+    assert ProxyLocation._to_deployment_mode("EveryNode") == DeploymentMode.EveryNode
+
+    with pytest.raises(ValueError):
+        ProxyLocation._to_deployment_mode("Unknown")
+
+    with pytest.raises(TypeError):
+        ProxyLocation._to_deployment_mode({"some_other_obj"})
 
 
 if __name__ == "__main__":
