+++ python/ray/serve/tests/test_gcs_failure.py
@@ -10,7 +10,7 @@ from ray.tests.conftest import external_redis  # noqa: F401
 from ray._private.test_utils import wait_for_condition
 
 from ray import serve
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from ray.serve._private.constants import (
     SERVE_DEFAULT_APP_NAME,
 )
@@ -27,7 +27,7 @@ def serve_ha(external_redis, monkeypatch):  # noqa: F811
         _system_config={"metrics_report_interval_ms": 1000, "task_retry_delay_ms": 50},
     )
     serve.start()
-    yield (address_info, get_global_client())
+    yield (address_info, _get_global_client())
     ray.shutdown()
