+++ python/ray/serve/tests/test_standalone.py
@@ -22,7 +22,7 @@ from ray._private.test_utils import (
     wait_for_condition,
 )
 from ray.cluster_utils import Cluster, cluster_not_supported
-from ray.serve.config import DeploymentMode, HTTPOptions
+from ray.serve.config import DeploymentMode, HTTPOptions, ProxyLocation
 from ray.serve._private.constants import (
     SERVE_DEFAULT_APP_NAME,
     SERVE_NAMESPACE,
@@ -30,7 +30,7 @@ from ray.serve._private.constants import (
     SERVE_ROOT_URL_ENV_KEY,
 )
 from ray._raylet import GcsClient
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from ray.serve.exceptions import RayServeException
 from ray.serve.generated.serve_pb2 import ActorNameList
 from ray.serve._private.http_util import set_socket_reuse_port
@@ -72,7 +72,7 @@ def lower_slow_startup_threshold_and_reset():
 
     ray.init(num_cpus=2)
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     yield client
 
@@ -649,7 +649,7 @@ def test_fixed_number_proxies(monkeypatch, ray_cluster):
     serve.run(A.bind())
 
     # Only the controller and two http proxy should be started.
-    controller_handle = get_global_client()._controller
+    controller_handle = _get_global_client()._controller
     wait_for_condition(
         lambda: len(ray.get(controller_handle.get_http_proxies.remote())) == 2
     )
@@ -666,7 +666,7 @@ def test_fixed_number_proxies(monkeypatch, ray_cluster):
 def test_serve_shutdown(ray_shutdown):
     ray.init(namespace="serve")
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     @serve.deployment
     class A:
@@ -679,7 +679,7 @@ def test_serve_shutdown(ray_shutdown):
 
     serve.shutdown()
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     assert len(client.list_deployments()) == 0
 
@@ -761,7 +761,7 @@ def test_recovering_controller_no_redeploy():
     ray_context = ray.init(namespace="x")
     address = ray_context.address_info["address"]
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     @serve.deployment
     def f():
@@ -868,7 +868,7 @@ def test_run_graph_task_uses_zero_cpus():
 
     ray.init(num_cpus=2)
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     config = {"import_path": "ray.serve.tests.test_standalone.WaiterNode"}
     config = ServeApplicationSchema.parse_obj(config)
@@ -905,32 +905,32 @@ def test_run_graph_task_uses_zero_cpus():
             "expected": HTTPOptions(location=DeploymentMode.NoServer),
         },
         {
-            "proxy_location": "NoServer",
+            "proxy_location": "Disabled",
             "http_options": None,
             "expected": HTTPOptions(location=DeploymentMode.NoServer),
         },
         {
-            "proxy_location": "NoServer",
+            "proxy_location": "Disabled",
             "http_options": {},
             "expected": HTTPOptions(location=DeploymentMode.NoServer),
         },
         {
-            "proxy_location": "NoServer",
+            "proxy_location": "Disabled",
             "http_options": HTTPOptions(host="foobar"),
             "expected": HTTPOptions(location=DeploymentMode.NoServer, host="foobar"),
         },
         {
-            "proxy_location": "NoServer",
+            "proxy_location": "Disabled",
             "http_options": {"host": "foobar"},
             "expected": HTTPOptions(location=DeploymentMode.NoServer, host="foobar"),
         },
         {
-            "proxy_location": "NoServer",
+            "proxy_location": "Disabled",
             "http_options": {"location": "HeadOnly"},
             "expected": HTTPOptions(location=DeploymentMode.NoServer),
         },
         {
-            "proxy_location": DeploymentMode.NoServer,
+            "proxy_location": ProxyLocation.Disabled,
             "http_options": HTTPOptions(location=DeploymentMode.HeadOnly),
             "expected": HTTPOptions(location=DeploymentMode.NoServer),
         },
@@ -939,7 +939,7 @@ def test_run_graph_task_uses_zero_cpus():
 def test_serve_start_proxy_location(ray_shutdown, options):
     expected_options = options.pop("expected")
     serve.start(**options)
-    client = get_global_client()
+    client = _get_global_client()
     assert ray.get(client._controller.get_http_config.remote()) == expected_options
