+++ python/ray/serve/tests/test_standalone_2.py
@@ -16,7 +16,7 @@ from ray import serve
 from ray._private.test_utils import wait_for_condition
 from ray.exceptions import RayActorError
 from ray.serve._private.constants import SERVE_NAMESPACE, SERVE_DEFAULT_APP_NAME
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from ray.tests.conftest import call_ray_stop_only  # noqa: F401
 
 
@@ -162,7 +162,7 @@ def test_refresh_controller_after_death(shutdown_ray_and_serve, detached):
     serve.shutdown()  # Ensure serve isn't running before beginning the test
     serve.start(detached=detached)
 
-    old_handle = get_global_client()._controller
+    old_handle = _get_global_client()._controller
     ray.kill(old_handle, no_restart=True)
 
     def controller_died(handle):
@@ -177,7 +177,7 @@ def test_refresh_controller_after_death(shutdown_ray_and_serve, detached):
     # Call start again to refresh handle
     serve.start(detached=detached)
 
-    new_handle = get_global_client()._controller
+    new_handle = _get_global_client()._controller
     assert new_handle is not old_handle
 
     # Health check should not error
@@ -193,7 +193,7 @@ def test_get_serve_status(shutdown_ray_and_serve):
 
     serve.run(f.bind())
 
-    client = get_global_client()
+    client = _get_global_client()
     status_info_1 = client.get_serve_status()
     assert status_info_1.app_status.status == "RUNNING"
     assert status_info_1.deployment_statuses[0].name == "f"
@@ -251,7 +251,7 @@ def test_controller_deserialization_deployment_def(
 def test_controller_deserialization_args_and_kwargs(shutdown_ray_and_serve):
     """Ensures init_args and init_kwargs stay serialized in controller."""
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     class PidBasedString(str):
         pass
@@ -292,7 +292,7 @@ def test_controller_recover_and_delete(shutdown_ray_and_serve):
 
     ray_context = ray.init()
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
 
     num_replicas = 10
