+++ python/ray/serve/tests/test_standalone_3.py
@@ -22,7 +22,7 @@ from ray.serve._private.constants import (
     RAY_SERVE_ENABLE_EXPERIMENTAL_STREAMING,
     SERVE_DEFAULT_APP_NAME,
 )
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from ray.serve.schema import ServeInstanceDetails
 from ray.serve._private.utils import get_head_node_id
 from ray.serve._private.common import ProxyStatus
@@ -290,7 +290,7 @@ def test_handle_early_detect_failure(shutdown_ray):
     pids = ray.get([handle.remote() for _ in range(2)])
     assert len(set(pids)) == 2
 
-    client = get_global_client()
+    client = _get_global_client()
     # Kill the controller so that the replicas membership won't be updated
     # through controller health check + long polling.
     ray.kill(client._controller, no_restart=True)
@@ -357,7 +357,7 @@ def test_autoscaler_shutdown_node_http_everynode(
         )
         == 2
     )
-    client = get_global_client()
+    client = _get_global_client()
     serve_details = ServeInstanceDetails(
         **ray.get(client._controller.get_serve_instance_details.remote())
     )
@@ -400,7 +400,7 @@ def test_drain_and_undrain_http_proxy_actors(
     wait_for_condition(lambda: len(ray._private.state.actors()) == 6)
     assert len(ray.nodes()) == 3
 
-    client = get_global_client()
+    client = _get_global_client()
     serve_details = ServeInstanceDetails(
         **ray.get(client._controller.get_serve_instance_details.remote())
     )
@@ -596,7 +596,7 @@ def test_controller_shutdown_gracefully(
     assert len(ray.nodes()) == 2
 
     # Call `graceful_shutdown()` on the controller, so it will start shutdown.
-    client = get_global_client()
+    client = _get_global_client()
     if wait_for_controller_shutdown:
         # Waiting for controller shutdown will throw RayActorError when the controller
         # killed itself.
@@ -647,7 +647,7 @@ def test_client_shutdown_gracefully_when_timeout(
 
     # Ensure client times out if the controller does not shutdown within timeout.
     timeout_s = 0.0
-    client = get_global_client()
+    client = _get_global_client()
     client.shutdown(timeout_s=timeout_s)
     assert (
         f"Controller failed to shut down within {timeout_s}s. "
