+++ python/ray/serve/tests/test_telemetry.py
@@ -12,7 +12,7 @@ from ray.dag.input_node import InputNode
 from ray._private.test_utils import wait_for_condition
 
 from ray import serve
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from ray.serve.drivers import DefaultgRPCDriver, DAGDriver
 from ray.serve.http_adapters import json_request
 from ray.serve.schema import ServeDeploySchema
@@ -223,7 +223,7 @@ def test_rest_api(manage_ray_with_telemetry, tmp_dir, version):
 
     subprocess.check_output(["serve", "deploy", config_file_path])
 
-    client = get_global_client()
+    client = _get_global_client()
     if version == "v2":
         # Make sure the applications are RUNNING.
         wait_for_condition(
@@ -352,7 +352,7 @@ def test_lightweight_config_options(
 
     # Deploy first config
     serve.start()
-    client = get_global_client()
+    client = _get_global_client()
     client.deploy_apps(ServeDeploySchema(**config))
     wait_for_condition(
         lambda: serve.status().applications["receiver_app"].status
@@ -589,7 +589,7 @@ def test_multiplexed_detect(manage_ray_with_telemetry):
     report = ray.get(storage_handle.get_report.remote())
     assert ServeUsageTag.MULTIPLEXED_API_USED.get_value_from_report(report) is None
 
-    client = get_global_client()
+    client = _get_global_client()
     wait_for_condition(
         lambda: client.get_serve_status("app").app_status.status
         == ApplicationStatus.RUNNING,
