+++ release/serve_tests/workloads/serve_handle_long_chain.py
@@ -19,7 +19,7 @@ from typing import Optional
 
 import ray
 from ray import serve
-from ray.serve.context import get_global_client
+from ray.serve.context import _get_global_client
 from serve_test_cluster_utils import (
     setup_local_single_node_cluster,
     setup_anyscale_cluster,
@@ -70,7 +70,7 @@ def construct_long_chain_graph_with_pure_handle(
         Node.options(name=str(id)).deploy(
             id, prev_handle, init_delay_secs, compute_delay_secs, sync_handle
         )
-        prev_handle = get_global_client().get_handle(
+        prev_handle = _get_global_client().get_handle(
             str(id), app_name="", sync=sync_handle
         )
     return prev_handle
